move parseMapData back
