+++ src/components/MapView/Interactions.js
@@ -1,6 +1,3 @@
-import { KML_FIELDS, PLACE_FIELDS } from "@data/dbUtils";
-import { processCoordinates, KML_TYPES } from '@utils/kmlUtils';
-
 export const isMapItemSelected = (id, selectedMapItem) => !!(selectedMapItem && selectedMapItem.rowid === id);
 
 export const openMarker = (selectedMarker, markersRef) => {
@@ -17,40 +14,5 @@ export const centerMap = (region, mapRef) => {
   mapRef.current.animateToRegion(region, 300);
 } 
 
-// TODO tranfer this to a context ?
-export const parseMapData = (mapData=[], placesData=[]) => {
-  const markerData = [],
-    polygonData = [],
-    polylineData = [];
 
-  mapData.forEach((data) => {
-    
-    // see if there is a place description entry for our thing
-    // places and kml have filenames in common
-    const placeData = (placesData.find((place) => place[PLACE_FIELDS.filename] === data[KML_FIELDS.filename]))
-      || {};
-    
-    const mapObject = { 
-      [KML_FIELDS.filename]: data[KML_FIELDS.filename],
-      coordinates: processCoordinates(data[KML_FIELDS.coordinates]),
-      rowid: data.rowid, // TODO improve rowid hardcode
-      type: data[KML_FIELDS.type],
-      placeData
-    };
-
-    if (data[KML_FIELDS.type] === KML_TYPES.Polygon) {
-      polygonData.push(mapObject);
-    } else if (data[KML_FIELDS.type] === KML_TYPES.Point) {
-      //const markerObject = { ...mapObject, ...placeData }; 
-      markerData.push(mapObject);
-    } else {
-      polylineData.push(mapObject); // KML_TYPE Polyline and Track
-    }
-  });
-  return {
-    markers: markerData,
-    polygons: polygonData,
-    polylines: polylineData
-  };
-}
